Refacto: Rename containers files and components
