+++ dataprep-webapp/src/app/components/widgets-containers/layout/layout.scss
