+++ dataprep-webapp/src/app/index.scss
@@ -105,5 +105,5 @@ $navbar-margin-bottom: 0;
 @import "components/widgets/charts/horizontal-barchart/horizontal-barchart.scss";
 @import "components/widgets/charts/range-slider/range-slider.scss";
 @import "components/widgets/charts/vertical-barchart/vertical-barchart.scss";
-@import "components/widgets-containers/app-layout/app-layout.scss";
 @import "components/widgets-containers/app-header-bar/app-header-bar.scss";
+@import "components/widgets-containers/layout/layout.scss";
